Use Hurl.library.Models instead - These models are now exclusively for Ser/De only
