+++ Source/Hurl.BrowserSelector/Helpers/IconExtractor.cs
@@ -1,4 +1,5 @@
-﻿using System.ComponentModel;
+﻿using System;
+using System.ComponentModel;
 using System.Diagnostics;
 using System.Drawing;
 using System.Runtime.InteropServices;
@@ -7,7 +8,7 @@ using System.Windows.Interop;
 using System.Windows.Media;
 using System.Windows.Media.Imaging;
 
-namespace Hurl.Library;
+namespace Hurl.BrowserSelector.Helpers;
 
 public static class IconExtractor
 {
