+++ Source/Hurl.Library/GetBrowsers.cs
@@ -30,7 +30,7 @@ public static class GetBrowsers
 
                 if (Name is string _Name && ExePath is string _ExePath)
                 {
-                    Browser b = new() { Name = Name, ExePath = ExePath };
+                    Browser b = new(Name, ExePath);
                     browsers.Add(b);
                 }
             }
