+++ Source/Hurl.Selector/Hurl.Selector.csproj
@@ -29,12 +29,12 @@
 
   <ItemGroup>
     <PackageReference Include="CommunityToolkit.Mvvm" Version="8.4.2" />
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="10.0.5" />
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="10.0.7" />
 		<PackageReference Include="Microsoft.Windows.CsWinRT" Version="2.2.0" />
-    <PackageReference Include="Microsoft.Windows.SDK.BuildTools" Version="10.0.28000.1721" />
-    <PackageReference Include="Microsoft.WindowsAppSDK" Version="1.8.260317003" GeneratePathProperty="true" />
-    <PackageReference Include="System.CommandLine" Version="2.0.5" />
-		<PackageReference Include="System.Drawing.Common" Version="10.0.5" />
+    <PackageReference Include="Microsoft.Windows.SDK.BuildTools" Version="10.0.28000.1839" />
+    <PackageReference Include="Microsoft.WindowsAppSDK" Version="2.0.1" GeneratePathProperty="true" />
+    <PackageReference Include="System.CommandLine" Version="2.0.7" />
+		<PackageReference Include="System.Drawing.Common" Version="10.0.7" />
     <PackageReference Include="WinUIEx" Version="2.9.0" />
     <Manifest Include="$(ApplicationManifest)" />
   </ItemGroup>
